qlog_dancer

Module plots

Modules

colors

congestion_control

conn_flow_control

conn_overview

minmax

packet_received

packet_sent

pending

Pending charts A single image file containing a view of pending vs. in-flight requests

rtt

stream_multiplex

Multiplexing charts A single image file containing various renderings of stream and H3 DATA

stream_sparks

Spark charts These are a single image file containing a grid of mini plots. Each plot represents a single stream, so it is easier to see how they all compare at a glance.

Structs

AreaMargin

ChartMargin

ChartSize

ClampParams

PlotParameters

Enums

ChartOutputType

Functions

chart_label_style

chart_subtitle_style

chart_title_style

make_chart_bitmap_area