qlog_dancer::plots

Module stream_sparks

Spark charts These are a single image file containing a grid of mini plots. Each plot represents a single stream, so it is easier to see how they all compare at a glance.

Structs

SparkPlotsParams

Enums

TransmissionType

Functions

plot_sparks